Migrate from CommonJS modules to ES2015 modules
