+++ README.md
@@ -65,20 +65,23 @@ performance, small size and all the features listed below.
 ```javascript
 var snabbdom = require('snabbdom');
 var patch = snabbdom.init([ // Init patch function with chosen modules
-  require('snabbdom/modules/class'), // makes it easy to toggle classes
-  require('snabbdom/modules/props'), // for setting properties on DOM elements
-  require('snabbdom/modules/style'), // handles styling on elements with support for animations
-  require('snabbdom/modules/eventlisteners'), // attaches event listeners
+  require('snabbdom/modules/class').default, // makes it easy to toggle classes
+  require('snabbdom/modules/props').default, // for setting properties on DOM elements
+  require('snabbdom/modules/style').default, // handles styling on elements with support for animations
+  require('snabbdom/modules/eventlisteners').default, // attaches event listeners
 ]);
-var h = require('snabbdom/h'); // helper function for creating vnodes
+var h = require('snabbdom/h').default; // helper function for creating vnodes
+
+var container = document.getElementById('container');
+
 var vnode = h('div#container.two.classes', {on: {click: someFn}}, [
   h('span', {style: {fontWeight: 'bold'}}, 'This is bold'),
   ' and this is just normal text',
   h('a', {props: {href: '/foo'}}, 'I\'ll take you places!')
 ]);
-var container = document.getElementById('container');
 // Patch into empty DOM element – this modifies the DOM as a side effect
 patch(container, vnode);
+
 var newVnode = h('div#container.two.classes', {on: {click: anotherEventHandler}}, [
   h('span', {style: {fontWeight: 'normal', fontStyle: 'italic'}}, 'This is now italic type'),
   ' and this is still just normal text',
@@ -108,8 +111,8 @@ specified set of modules.
 
 ```javascript
 var patch = snabbdom.init([
-  require('snabbdom/modules/class'),
-  require('snabbdom/modules/style'),
+  require('snabbdom/modules/class').default,
+  require('snabbdom/modules/style').default,
 ]);
 ```
 
@@ -141,7 +144,7 @@ tag/selector as a string, an optional data object and an optional string or
 array of children.
 
 ```javascript
-var h = require('snabbdom/h');
+var h = require('snabbdom/h').default;
 var vnode = h('div', {style: {color: '#000'}}, [
   h('h1', 'Headline'),
   h('p', 'A paragraph'),
@@ -478,7 +481,7 @@ h('text', {
       y: yPos,
       dy: "5",
       class: 'text_class'
-    }}, 
+    }},
   text
 );
 ```
