+++ examples/reorder-animation/script.js
@@ -1,11 +1,11 @@
 var snabbdom = require('../../snabbdom.js');
 var patch = snabbdom.init([
-  require('../../modules/class'),
-  require('../../modules/props'),
-  require('../../modules/style'),
-  require('../../modules/eventlisteners'),
+  require('../../modules/class').default,
+  require('../../modules/props').default,
+  require('../../modules/style').default,
+  require('../../modules/eventlisteners').default,
 ]);
-var h = require('../../h.js');
+var h = require('../../h.js').default;
 
 var vnode;
