+++ src/is.ts
@@ -1,6 +1,4 @@
-export = {
-  array: Array.isArray,
-  primitive: function primitive(s: any): boolean {
-    return typeof s === 'string' || typeof s === 'number';
-  },
+export const array = Array.isArray;
+export function primitive(s: any): boolean {
+  return typeof s === 'string' || typeof s === 'number';
 };
