+++ src/modules/module.d.ts
@@ -1,0 +1,10 @@
+import {PreHook, CreateHook, UpdateHook, DestroyHook, RemoveHook, PostHook} from '../hooks';
+
+export interface Module {
+  pre?: PreHook;
+  create?: CreateHook;
+  update?: UpdateHook;
+  destroy?: DestroyHook;
+  remove?: RemoveHook;
+  post?: PostHook;
+}
