+++ src/snabbdom.bundle.ts
@@ -1,16 +1,16 @@
-import snabbdom = require('./snabbdom');
-import attributesModule = require('./modules/attributes'); // for setting attributes on DOM elements
-import classModule = require('./modules/class'); // makes it easy to toggle classes
-import propsModule = require('./modules/props'); // for setting properties on DOM elements
-import styleModule = require('./modules/style'); // handles styling on elements with support for animations
-import eventListenersModule = require('./modules/eventlisteners'); // attaches event listeners
-import h = require('./h'); // helper function for creating vnodes
-var patch = snabbdom.init([ // Init patch function with choosen modules
+import {init} from './snabbdom';
+import {attributesModule} from './modules/attributes'; // for setting attributes on DOM elements
+import {classModule} from './modules/class'; // makes it easy to toggle classes
+import {propsModule} from './modules/props'; // for setting properties on DOM elements
+import {styleModule} from './modules/style'; // handles styling on elements with support for animations
+import {eventListenersModule} from './modules/eventlisteners'; // attaches event listeners
+import {h} from './h'; // helper function for creating vnodes
+var patch = init([ // Init patch function with choosen modules
   attributesModule,
   classModule,
   propsModule,
   styleModule,
   eventListenersModule
 ]) as (oldVNode: any, vnode: any) => any;
-
-export = { patch, h: h as any };
+export const snabbdomBundle = { patch, h: h as any };
+export default snabbdomBundle;
