+++ src/vnode.ts
@@ -1,6 +1,34 @@
-import {VNode} from './interfaces';
+import {Hooks} from './hooks';
 
-function vnode(sel: string,
+export interface VNode {
+  sel: string | undefined;
+  data: VNodeData | undefined;
+  children: Array<VNode | string> | undefined;
+  elm: Element | Text | undefined;
+  text: string | undefined;
+  key: string | number | undefined;
+}
+
+export interface VNodeData {
+  // modules - use any because Object type is useless
+  props?: any;
+  attrs?: any;
+  class?: any;
+  style?: any;
+  dataset?: any;
+  on?: any;
+  hero?: any;
+  attachData?: any;
+  [key: string]: any; // for any other 3rd party module
+  // end of modules
+  hook?: Hooks;
+  key?: string | number;
+  ns?: string; // for SVGs
+  fn?: () => VNode; // for thunks
+  args?: Array<any>; // for thunks
+}
+
+export function vnode(sel: string,
                data: any | undefined,
                children: Array<VNode | string> | undefined,
                text: string | undefined,
@@ -10,4 +38,4 @@ function vnode(sel: string,
           text: text, elm: elm, key: key};
 }
 
-export = vnode;
+export default vnode;
