+++ tsconfig.json
@@ -22,10 +22,11 @@
         "src/modules/eventlisteners.ts",
         "src/modules/hero.ts",
         "src/modules/props.ts",
+        "src/modules/module.d.ts",
         "src/modules/style.ts",
         "src/h.ts",
         "src/htmldomapi.ts",
-        "src/interfaces.d.ts",
+        "src/hooks.d.ts",
         "src/is.ts",
         "src/snabbdom.bundle.ts",
         "src/snabbdom.ts",
